+++ src/main/java/shop/zip/travel/domain/base/BaseTimeEntity.java
@@ -1,4 +1,4 @@
-package shop.zip.travel.global.entity;
+package shop.zip.travel.domain.base;
 
 import java.time.LocalDateTime;
