+++ src/main/java/shop/zip/travel/domain/image/entity/Image.java
@@ -1,4 +1,4 @@
-package shop.zip.travel.domain.travelog.entity;
+package shop.zip.travel.domain.image.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -8,7 +8,8 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import shop.zip.travel.global.entity.BaseImage;
+import shop.zip.travel.domain.post.travelog.entity.SubTravelogue;
+import shop.zip.travel.domain.base.BaseImage;
 
 @Entity
 public class Image implements BaseImage {
