+++ src/test/java/shop/zip/travel/domain/travelog/entity/PeriodTest.java
@@ -9,6 +9,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
+import shop.zip.travel.domain.post.travelog.entity.type.Period;
+
 class PeriodTest {
 
 	@DisplayName("startDate 가 endDate 보다 나중일 경우 IllegalArgumentException 을 던진다")
